*api continuous.........

+++ app/src/main/java/com/dukan/dukkan/pojo/LatestOffer.java
@@ -36,6 +36,6 @@ public class LatestOffer {
     public IsCart isCart;
     @SerializedName("rate")
     @Expose
-    public Integer rate;
+    public float rate;
 
 }
