+++ app/src/main/java/com/dukan/dukkan/pojo/NewProduct.java
@@ -36,6 +36,6 @@ public class NewProduct {
     public IsCart isCart;
     @SerializedName("rate")
     @Expose
-    public Integer rate;
+    public float rate;
 
 }
