+++ app/src/main/java/com/dukan/dukkan/pojo/Store.java
@@ -71,5 +71,5 @@ public class Store {
     public Integer customersCount;
     @SerializedName("rate")
     @Expose
-    public Integer rate;
+    public float rate;
 }
